fixed shell script bug and removed redundant logging in fault injection enable/disable code

+++ server/fault-injection/disable_abort_fault.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 set -e
 
-rm /srv/runtime/v1/envoy/fault/http/abort/abort_percent
-rm /srv/runtime/v1/envoy/fault/http/abort/http_status
+[ -e /srv/runtime/v1/envoy/fault/http/abort/abort_percent ] && rm /srv/runtime/v1/envoy/fault/http/abort/abort_percent
+[ -e /srv/runtime/v1/envoy/fault/http/abort/http_status ] && rm /srv/runtime/v1/envoy/fault/http/abort/http_status
 
 pushd /srv/runtime
 ln -s /srv/runtime/v1 new && mv -Tf new current
